ultimos cambios

+++ src/screens/Administrador/Programas.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 
 // CSS
-import { Row, Col, Table, Button } from "antd";
+import { Row, Col, Table, Button, Modal } from "antd";
 import { DeleteOutlined, EditOutlined, PlusOutlined, LeftCircleOutlined } from '@ant-design/icons';
 import { useHistory } from "react-router-dom";
 
@@ -32,8 +32,8 @@ export default function Programas() {
         },
         {
             title: "Jornada",
-            dataIndex: "jornada",
-            key: "jornada",
+            dataIndex: "jornada1",
+            key: "jornada1",
         },
         {
             title: "Facultad",
@@ -82,15 +82,20 @@ export default function Programas() {
     }, [listactualizada])
 
     const handleDelete = id => {
-        const requestInit = {
-            method: 'DELETE'
-        }
-        fetch('https://app-gestion-aunar.herokuapp.com/programa/' + id, requestInit)
-            .then(res => res.text())
-            .then(res => console.log(res))
-        alert('Eliminado Exitosamente');
-
-        setListaactualizada(true);
+        Modal.confirm({
+            title: "Estas seguro de eliminar el Docente",
+            okText: "Si",
+            okType: "danger",
+            onOk: () => {
+                const requestInit = {
+                    method: 'DELETE'
+                }
+                fetch('https://app-gestion-aunar.herokuapp.com/programa/' + id, requestInit)
+                    .then(res => res.text())
+                    .then(res => console.log(res))
+                setListaactualizada(true);
+            }
+        });
     }
 
     const handleUpdate = id => {
@@ -113,7 +118,7 @@ export default function Programas() {
                 <Row className="box-select-content border-radius-10 box-shadow" style={{ padding: "2%" }}>
                     <Row style={{ width: "100%", padding: "2%" }} className="d-flex justify-content-center">
                         <Col span={4}>
-                            <Button type="primary" shape="round" icon={<LeftCircleOutlined />} onClick={() => gotoScreen("/Administrador1/")} >
+                            <Button type="primary" shape="round" icon={<LeftCircleOutlined />} onClick={() => gotoScreen("/Administrador/")} >
                                 Salir
                             </Button>
                         </Col>
@@ -121,7 +126,7 @@ export default function Programas() {
                             <span className="titulos">Programas Académicos</span>
                         </Col>
                         <Col span={4}>
-                            <Button type="primary" shape="round" icon={<PlusOutlined />} onClick={() => gotoScreen("/Administrador1/CrearProgramas")} >
+                            <Button type="primary" shape="round" icon={<PlusOutlined />} onClick={() => gotoScreen("/Administrador/CrearProgramas")} >
                                 Crear Programa
                             </Button>
                         </Col>
