+++ src/screens/Administrador/Usuarios.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 // CSS
-import { Row, Col, Table, Button } from "antd";
+import { Row, Col, Table, Button, Modal } from "antd";
 import { DeleteOutlined, EditOutlined, PlusOutlined, LeftCircleOutlined } from '@ant-design/icons';
 import { useHistory } from "react-router-dom";
 
@@ -65,15 +65,20 @@ export default function Usuarios() {
   }, [listactualizada])
 
   const handleDelete = id => {
-    const requestInit = {
-      method: 'DELETE'
-    }
-    fetch('https://app-gestion-aunar.herokuapp.com/usuarios/' + id, requestInit)
-      .then(res => res.text())
-      .then(res => console.log(res))
-    alert('Eliminado Exitosamente');
-
-    setListaactualizada(true);
+    Modal.confirm({
+      title: "Estas seguro de eliminar el Usuario",
+      okText: "Si",
+      okType: "danger",
+      onOk: () => {
+        const requestInit = {
+          method: 'DELETE'
+        }
+        fetch('https://app-gestion-aunar.herokuapp.com/usuarios/' + id, requestInit)
+          .then(res => res.text())
+          .then(res => console.log(res))
+        setListaactualizada(true);
+      }
+    });
   }
 
   const handleUpdate = id => {
@@ -96,7 +101,7 @@ export default function Usuarios() {
         <Row className="box-select-content border-radius-10 box-shadow" style={{ padding: "2%" }}>
           <Row style={{ width: "100%", padding: "2%" }} className="d-flex justify-content-center">
             <Col span={4}>
-              <Button type="primary" shape="round" icon={<LeftCircleOutlined />} onClick={() => gotoScreen("/Administrador1/")}>
+              <Button type="primary" shape="round" icon={<LeftCircleOutlined />} onClick={() => gotoScreen("/Administrador/")}>
                 Salir
               </Button>
             </Col>
@@ -104,7 +109,7 @@ export default function Usuarios() {
               <span className="titulos">Usuarios</span>
             </Col>
             <Col span={4}>
-              <Button type="primary" shape="round" icon={<PlusOutlined />} onClick={() => gotoScreen("/Administrador1/CrearUsuarios")}>
+              <Button type="primary" shape="round" icon={<PlusOutlined />} onClick={() => gotoScreen("/Administrador/CrearUsuarios")}>
                 Crear usuario
               </Button>
             </Col>
